requirements.txt
adafruit-circuitpython-ads1x15==2.2.12
adafruit_extended_bus==1.0.2
luma.oled==3.8.1
pigpio==1.78
Pillow==9.0.1
